Refactor: Remove use cases and interactor from Movies module
This commit removes the following use cases and interactor from the Movies module:

- `GetDefaultEmptyStateUseCase`
- `GetDefaultMenuItemsUseCase`
- `GetMoviesFromFilterUseCase`
- `MoviesInteractor`

These have been replaced by the `MoviesHandler` class, which simplifies the domain layer. The models `FilterType`, `FilterMenuItem`, and `MoviesState` have been moved to the `model` package.
