--- src/feature/movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/movies/domain/usecase/getmovies/GetMoviesFromFilterUseCase.kt
@@ -1,9 +1,0 @@
-package com.gabrielbmoro.moviedb.movies.domain.usecase.getmovies
-
-import com.gabrielbmoro.moviedb.domain.entities.Movie
-import com.gabrielbmoro.moviedb.domain.usecases.UseCase
-import com.gabrielbmoro.moviedb.movies.domain.model.FilterType
-
-interface GetMoviesFromFilterUseCase : UseCase<GetMoviesFromFilterUseCase.Params, List<Movie>> {
-    data class Params(val filter: FilterType, val page: Int)
-}
