+++ src/feature/movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/movies/model/FilterMenuItem.kt
@@ -1,4 +1,4 @@
-package com.gabrielbmoro.moviedb.movies.domain.model
+package com.gabrielbmoro.moviedb.movies.model
 
 data class FilterMenuItem(
     val selected: Boolean,
