+++ src/feature/movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/movies/ui/widgets/FilterMenu.kt
@@ -2,7 +2,6 @@ package com.gabrielbmoro.moviedb.movies.ui.widgets
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
@@ -11,8 +10,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.gabrielbmoro.moviedb.movies.domain.model.FilterMenuItem
-import com.gabrielbmoro.moviedb.movies.domain.model.FilterType
+import com.gabrielbmoro.moviedb.movies.model.FilterMenuItem
+import com.gabrielbmoro.moviedb.movies.model.FilterType
 import moviedbapp.feature.movies.generated.resources.Res
 import moviedbapp.feature.movies.generated.resources.now_playing
 import moviedbapp.feature.movies.generated.resources.popular
